+++ src/main/java/dev/jaxydog/astral/content/item/group/AstralCycledItemGroup.java
@@ -19,95 +19,157 @@ import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
 import net.minecraft.text.Text;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.function.Supplier;
 
 /**
- * An extension of a {@link AstralItemGroup} that provides a cycling item icon.
+ * An extension of an {@link AstralItemGroup} that provides a cycling icon.
+ * <p>
+ * If you do not need more than one icon, it's recommended to use {@link AstralItemGroup} instead.
  *
  * @author Jaxydog
  */
 public class AstralCycledItemGroup extends AstralItemGroup {
 
-    private final List<Supplier<ItemStack>> stacks;
-    private final int interval;
+    /** A list containing all valid icons for this item group. */
+    protected final List<Supplier<ItemStack>> iconSuppliers = new ObjectArrayList<>();
+    /** A duration in ticks that determines the cycle speed of the icons. */
+    protected final int cycleInterval;
 
-    private int index = 0;
-    private float progress = 0F;
+    /** An index into the supplier list that determines the active icon. */
+    private int iconIndex = 0;
+    /** A counter that controls when to increment and cycle the active icon. */
+    private float cycleProgress = 0F;
 
-    private AstralCycledItemGroup(
-        String idPath,
+    /**
+     * Creates a new item group.
+     * <p>
+     * In most cases, you will want to instead use this class' builder.
+     *
+     * @param path The item group's identifier path.
+     * @param row The item group's row.
+     * @param column The item group's column.
+     * @param type The item group's type.
+     * @param displayName The item group's display name.
+     * @param entryCollector The item group's entry collector.
+     * @param icons The item group's preferred icons.
+     * @param cycleInterval The duration in ticks that determines the length of stay for one icon.
+     */
+    @SuppressWarnings("unused")
+    public AstralCycledItemGroup(
+        String path,
         Row row,
         int column,
         Type type,
         Text displayName,
         EntryCollector entryCollector,
-        List<Supplier<ItemStack>> stacks,
-        int interval
+        List<Supplier<ItemStack>> icons,
+        int cycleInterval
     ) {
-        super(idPath, row, column, type, displayName, () -> ItemStack.EMPTY, entryCollector);
+        super(path, row, column, type, displayName, () -> ItemStack.EMPTY, entryCollector);
 
-        this.stacks = stacks;
-        this.interval = interval;
+        this.iconSuppliers.addAll(icons);
+        this.cycleInterval = cycleInterval;
     }
 
+    /**
+     * Creates a new item group from a given source.
+     * <p>
+     * In most cases, you will want to instead use this class' builder.
+     *
+     * @param path The item group's identifier path.
+     * @param group The source item group.
+     * @param entryCollector The item group's entry collector.
+     * @param icons The item group's preferred icons.
+     * @param cycleInterval The duration in ticks that determines the length of stay for one icon.
+     */
     protected AstralCycledItemGroup(
-        String idPath, ItemGroup group, EntryCollector entryCollector, List<Supplier<ItemStack>> stacks, int interval
+        String path, ItemGroup group, EntryCollector entryCollector, List<Supplier<ItemStack>> icons, int cycleInterval
     ) {
-        this(idPath,
-            group.getRow(),
-            group.getColumn(),
-            group.getType(),
-            group.getDisplayName(),
-            entryCollector,
-            stacks,
-            interval
-        );
+        super(path, group, entryCollector);
+
+        this.iconSuppliers.addAll(icons);
+        this.cycleInterval = cycleInterval;
     }
 
-    public static Builder builder(String idPath) {
-        return new Builder(idPath);
+    /**
+     * Returns a new builder for a {@link AstralCycledItemGroup}.
+     *
+     * @param path The identifier path.
+     *
+     * @return A new builder.
+     */
+    @Contract("_ -> new")
+    public static @NotNull Builder builder(String path) {
+        return new Builder(path);
     }
 
     @Override
     public ItemStack astral$getIcon(float delta) {
-        this.progress += delta;
+        this.cycleProgress += delta;
 
-        if (this.progress >= this.interval) {
-            this.progress %= this.interval;
+        // If the progress exceeds the interval, wrap the progress and update the index.
+        if (this.cycleProgress >= (float) this.cycleInterval) {
+            this.cycleProgress %= (float) this.cycleInterval;
 
-            if (!this.stacks.isEmpty()) {
-                this.index += 1;
-                this.index %= this.stacks.size();
+            // Only update & wrap the index if there are at least two icons.
+            if (this.iconSuppliers.size() > 1) {
+                this.iconIndex += 1;
+                this.iconIndex %= this.iconSuppliers.size();
             }
         }
 
-        if (this.stacks.isEmpty()) {
-            return ItemStack.EMPTY;
+        // Only return an icon if the current index is within bounds.
+        if (this.iconSuppliers.size() > this.iconIndex) {
+            return this.iconSuppliers.get(this.iconIndex).get();
         } else {
-            return this.stacks.get(this.index).get();
+            return ItemStack.EMPTY;
         }
     }
 
+    /**
+     * Builds and constructs an instance of a new {@link AstralItemGroup}.
+     *
+     * @author Jaxydog
+     */
     public static class Builder extends AstralItemGroup.Builder {
 
-        private final List<Supplier<ItemStack>> stacks = new ObjectArrayList<>(1);
-        private int interval = 40;
+        /** A list of the currently added icon suppliers. */
+        // It's safe to assume that at least two icons will be added.
+        protected final List<Supplier<ItemStack>> iconSuppliers = new ObjectArrayList<>(2);
+        /** A duration in ticks that determines the cycle speed of the icons. */
+        protected int cycleInterval = 40;
 
-        public Builder(String idPath) {
-            super(idPath);
+        /**
+         * Creates a new builder instance.
+         * <p>
+         * This is only accessible through {@link #builder(String)} or subclasses.
+         *
+         * @param path The identifier path.
+         */
+        protected Builder(String path) {
+            super(path);
+        }
+
+        /**
+         * Sets the duration in ticks that determines the cycle speed of the icons.
+         *
+         * @param ticks The duration in ticks.
+         *
+         * @return The builder instance.
+         */
+        public Builder cycleInterval(int ticks) {
+            this.cycleInterval = ticks;
+
+            return this;
         }
 
         @Override
-        public Builder icon(Supplier<ItemStack> iconSupplier) {
-            this.stacks.add(iconSupplier);
-            return this;
-        }
-
-        public Builder interval(int ticks) {
-            this.interval = ticks;
-            return this;
+        public Builder displayName(Text displayName) {
+            return (Builder) super.displayName(displayName);
         }
 
         @Override
@@ -116,8 +178,10 @@ public class AstralCycledItemGroup extends AstralItemGroup {
         }
 
         @Override
-        public Builder special() {
-            return (Builder) super.special();
+        public Builder icon(Supplier<ItemStack> iconSupplier) {
+            this.iconSuppliers.add(iconSupplier);
+
+            return this;
         }
 
         @Override
@@ -131,8 +195,8 @@ public class AstralCycledItemGroup extends AstralItemGroup {
         }
 
         @Override
-        protected Builder type(Type type) {
-            return (Builder) super.type(type);
+        public Builder special() {
+            return (Builder) super.special();
         }
 
         @Override
@@ -141,14 +205,28 @@ public class AstralCycledItemGroup extends AstralItemGroup {
         }
 
         @Override
-        public AstralCycledItemGroup finish() {
-            final Text name = Text.translatable(Astral.getId(this.idPath).toTranslationKey("itemGroup"));
+        protected Builder type(Type type) {
+            return (Builder) super.type(type);
+        }
 
-            return new AstralCycledItemGroup(this.idPath,
-                super.displayName(name).build(),
+        @Override
+        public AstralCycledItemGroup build() {
+            final ItemGroup group;
+
+            // Only set a name if one has not yet been set.
+            if (!this.hasDisplayName) {
+                final Text name = Text.translatable(Astral.getId(this.path).toTranslationKey("itemGroup"));
+
+                group = super.displayName(name).build();
+            } else {
+                group = super.build();
+            }
+
+            return new AstralCycledItemGroup(this.path,
+                group,
                 this.entryCollector,
-                this.stacks,
-                this.interval
+                this.iconSuppliers,
+                this.cycleInterval
             );
         }