+++ src/main/java/dev/jaxydog/astral/content/item/group/AstralItemGroup.java
@@ -15,27 +15,45 @@
 package dev.jaxydog.astral.content.item.group;
 
 import dev.jaxydog.astral.Astral;
-import dev.jaxydog.astral.register.Registered;
+import dev.jaxydog.astral.register.Registered.Common;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.text.Text;
+import org.jetbrains.annotations.Contract;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.function.Supplier;
 
 /**
- * Simple extension to allow automatic registration of item groups.
+ * An extension of an {@link ItemGroup} that provides commonly used functionality.
+ * <p>
+ * This type is automatically registered.
  *
  * @author Jaxydog
  */
-public class AstralItemGroup extends ItemGroup implements Registered.Common {
+public class AstralItemGroup extends ItemGroup implements Common {
 
-    private final String idPath;
+    /** The item group's identifier path used within the registration system. */
+    private final String path;
 
-    protected AstralItemGroup(
-        String idPath,
+    /**
+     * Creates a new item group.
+     * <p>
+     * In most cases, you will want to instead use this class' builder.
+     *
+     * @param path The item group's identifier path.
+     * @param row The item group's row.
+     * @param column The item group's column.
+     * @param type The item group's type.
+     * @param displayName The item group's display name.
+     * @param iconSupplier The item group's preferred icon.
+     * @param entryCollector The item group's entry collector.
+     */
+    public AstralItemGroup(
+        String path,
         Row row,
         int column,
         Type type,
@@ -45,12 +63,21 @@ public class AstralItemGroup extends ItemGroup implements Registered.Common {
     ) {
         super(row, column, type, displayName, iconSupplier, entryCollector);
 
-        this.idPath = idPath;
+        this.path = path;
     }
 
-    protected AstralItemGroup(String idPath, ItemGroup group, EntryCollector entryCollector) {
+    /**
+     * Creates a new item group from a given source.
+     * <p>
+     * In most cases, you will want to instead use this class' builder.
+     *
+     * @param path The item group's identifier path.
+     * @param group The source item group.
+     * @param entryCollector The item group's entry collector.
+     */
+    protected AstralItemGroup(String path, ItemGroup group, EntryCollector entryCollector) {
         this(
-            idPath,
+            path,
             group.getRow(),
             group.getColumn(),
             group.getType(),
@@ -60,17 +87,32 @@ public class AstralItemGroup extends ItemGroup implements Registered.Common {
         );
     }
 
-    public static Builder builder(String idPath) {
-        return new Builder(idPath);
+    /**
+     * Returns a new builder for an {@link AstralItemGroup}.
+     *
+     * @param path The identifier path.
+     *
+     * @return A new builder.
+     */
+    @Contract("_ -> new")
+    public static @NotNull Builder builder(String path) {
+        return new Builder(path);
     }
 
+    /**
+     * Returns this item group's associated registry key.
+     * <p>
+     * If the item group has not been registered, this method will throw a run-time exception.
+     *
+     * @return The associated registry key.
+     */
     public RegistryKey<ItemGroup> getRegistryKey() {
         return Registries.ITEM_GROUP.getKey(this).orElseThrow();
     }
 
     @Override
-    public String getRegistryIdPath() {
-        return this.idPath;
+    public String getRegistryPath() {
+        return this.path;
     }
 
     @Override
@@ -78,20 +120,39 @@ public class AstralItemGroup extends ItemGroup implements Registered.Common {
         Registry.register(Registries.ITEM_GROUP, this.getRegistryId(), this);
     }
 
+    /**
+     * Builds and constructs an instance of a new {@link AstralItemGroup}.
+     *
+     * @author Jaxydog
+     */
     public static class Builder extends ItemGroup.Builder {
 
-        protected final String idPath;
-        protected EntryCollector entryCollector = (a, b) -> {};
+        /** The item group's identifier path used within the registration system. */
+        protected final String path;
+        /** The item group's inner entry collector. */
+        protected EntryCollector entryCollector = (context, entries) -> {};
+        /** Whether this item group builder has set a display name. */
+        protected boolean hasDisplayName = false;
 
-        public Builder(String idPath) {
+        /**
+         * Creates a new builder instance.
+         * <p>
+         * This is only accessible through {@link #builder(String)} or subclasses.
+         *
+         * @param path The identifier path.
+         */
+        protected Builder(String path) {
+            // Mimic's the default Fabric builder's constructor.
             super(null, -1);
 
-            this.idPath = idPath;
+            this.path = path;
         }
 
         @Override
-        public Builder icon(Supplier<ItemStack> iconSupplier) {
-            return (Builder) super.icon(iconSupplier);
+        public Builder displayName(Text displayName) {
+            this.hasDisplayName = true;
+
+            return (Builder) super.displayName(displayName);
         }
 
         @Override
@@ -102,8 +163,8 @@ public class AstralItemGroup extends ItemGroup implements Registered.Common {
         }
 
         @Override
-        public Builder special() {
-            return (Builder) super.special();
+        public Builder icon(Supplier<ItemStack> iconSupplier) {
+            return (Builder) super.icon(iconSupplier);
         }
 
         @Override
@@ -117,8 +178,8 @@ public class AstralItemGroup extends ItemGroup implements Registered.Common {
         }
 
         @Override
-        protected Builder type(Type type) {
-            return (Builder) super.type(type);
+        public Builder special() {
+            return (Builder) super.special();
         }
 
         @Override
@@ -126,10 +187,25 @@ public class AstralItemGroup extends ItemGroup implements Registered.Common {
             return (Builder) super.texture(texture);
         }
 
-        public AstralItemGroup finish() {
-            final Text name = Text.translatable(Astral.getId(this.idPath).toTranslationKey("itemGroup"));
+        @Override
+        protected Builder type(Type type) {
+            return (Builder) super.type(type);
+        }
 
-            return new AstralItemGroup(this.idPath, super.displayName(name).build(), this.entryCollector);
+        @Override
+        public AstralItemGroup build() {
+            final ItemGroup group;
+
+            // Only set a name if one has not yet been set.
+            if (!this.hasDisplayName) {
+                final Text name = Text.translatable(Astral.getId(this.path).toTranslationKey("itemGroup"));
+
+                group = super.displayName(name).build();
+            } else {
+                group = super.build();
+            }
+
+            return new AstralItemGroup(this.path, group, this.entryCollector);
         }
 
     }